+++ api/index.tsx
@@ -1,2 +1,3 @@
 export * from './stockApi';
 export * from './types';
+export * from './authApi';