+++ src/components/SavedSims.tsx
@@ -1,0 +1,9 @@
+const SavedSims = () => {
+    return(
+        <>
+            <h1> SAVED SIMULATIONS</h1>   
+        </>
+  );
+};
+
+export default SavedSims;