+++ store/atoms/index.ts
@@ -1,2 +1,3 @@
 export * from './paramsAtom'
 export * from './stocksAtom'
+export * from './authAtom'